Moved the randomQuotes logic into another file
